Refactor package and setup.py .
