+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include personal_pre_commit_hooks/configs/*
