+++ Pipfile
@@ -14,6 +14,7 @@ yapf = "*"
 [dev-packages]
 personal-pre-commit-hooks = { path = ".", editable=true }
 pytest = "*"
+toml = "*"
 
 [requires]
-python_version = "3.10"
+python_version = "3.8"
