+++ personal_pre_commit_hooks/configs/pylintrc
