+++ personal_pre_commit_hooks/hooks/symbolic_links.sh
