+++ personal_pre_commit_hooks/hooks/utils.sh
